impl data-tapes + compiler packages

+++ src/main/java/com/tregouet/context_interpreter/data_types/representation/op/IDSOperator.java
@@ -4,7 +4,7 @@ public interface IDSOperator {
 	
 	int getContextStateID();
 	
-	float getCost();
+	double getCost();
 	
 	int getID();
